[ML] refactor internal datafeed management
