+++ x-pack/plugin/ml/src/main/java/org/elasticsearch/xpack/ml/action/TransportDeleteDatafeedAction.java
@@ -19,7 +19,6 @@ import org.elasticsearch.cluster.metadata.IndexNameExpressionResolver;
 import org.elasticsearch.cluster.service.ClusterService;
 import org.elasticsearch.common.inject.Inject;
 import org.elasticsearch.common.settings.Settings;
-import org.elasticsearch.common.xcontent.NamedXContentRegistry;
 import org.elasticsearch.persistent.PersistentTasksCustomMetadata;
 import org.elasticsearch.persistent.PersistentTasksService;
 import org.elasticsearch.tasks.Task;
@@ -28,12 +27,9 @@ import org.elasticsearch.transport.TransportService;
 import org.elasticsearch.xpack.core.ml.MlTasks;
 import org.elasticsearch.xpack.core.ml.action.DeleteDatafeedAction;
 import org.elasticsearch.xpack.core.ml.action.IsolateDatafeedAction;
-import org.elasticsearch.xpack.core.ml.datafeed.DatafeedState;
-import org.elasticsearch.xpack.core.ml.job.messages.Messages;
 import org.elasticsearch.xpack.core.ml.utils.ExceptionsHelper;
 import org.elasticsearch.xpack.ml.MlConfigMigrationEligibilityCheck;
-import org.elasticsearch.xpack.ml.datafeed.persistence.DatafeedConfigProvider;
-import org.elasticsearch.xpack.ml.job.persistence.JobDataDeleter;
+import org.elasticsearch.xpack.ml.datafeed.DatafeedManager;
 
 import static org.elasticsearch.xpack.core.ClientHelper.ML_ORIGIN;
 import static org.elasticsearch.xpack.core.ClientHelper.executeAsyncWithOrigin;
@@ -41,8 +37,7 @@ import static org.elasticsearch.xpack.core.ClientHelper.executeAsyncWithOrigin;
 public class TransportDeleteDatafeedAction extends AcknowledgedTransportMasterNodeAction<DeleteDatafeedAction.Request> {
 
     private final Client client;
-    private final DatafeedConfigProvider datafeedConfigProvider;
-    private final ClusterService clusterService;
+    private final DatafeedManager datafeedManager;
     private final PersistentTasksService persistentTasksService;
     private final MlConfigMigrationEligibilityCheck migrationEligibilityCheck;
 
@@ -51,14 +46,13 @@ public class TransportDeleteDatafeedAction extends AcknowledgedTransportMasterNo
                                          ThreadPool threadPool, ActionFilters actionFilters,
                                          IndexNameExpressionResolver indexNameExpressionResolver,
                                          Client client, PersistentTasksService persistentTasksService,
-                                         NamedXContentRegistry xContentRegistry) {
+                                         DatafeedManager datafeedManager) {
         super(DeleteDatafeedAction.NAME, transportService, clusterService, threadPool, actionFilters,
                 DeleteDatafeedAction.Request::new, indexNameExpressionResolver, ThreadPool.Names.SAME);
         this.client = client;
-        this.datafeedConfigProvider = new DatafeedConfigProvider(client, xContentRegistry);
         this.persistentTasksService = persistentTasksService;
-        this.clusterService = clusterService;
         this.migrationEligibilityCheck = new MlConfigMigrationEligibilityCheck(settings, clusterService);
+        this.datafeedManager = datafeedManager;
     }
 
     @Override
@@ -73,14 +67,15 @@ public class TransportDeleteDatafeedAction extends AcknowledgedTransportMasterNo
         if (request.isForce()) {
             forceDeleteDatafeed(request, state, listener);
         } else {
-            deleteDatafeedConfig(request, listener);
+            datafeedManager.deleteDatafeed(request, state, listener);
         }
     }
 
     private void forceDeleteDatafeed(DeleteDatafeedAction.Request request, ClusterState state,
                                      ActionListener<AcknowledgedResponse> listener) {
         ActionListener<Boolean> finalListener = ActionListener.wrap(
-                response -> deleteDatafeedConfig(request, listener),
+                // use clusterService.state() here so that the updated state without the task is available
+                response -> datafeedManager.deleteDatafeed(request, clusterService.state(), listener),
                 listener::onFailure
         );
 
@@ -117,37 +112,6 @@ public class TransportDeleteDatafeedAction extends AcknowledgedTransportMasterNo
                         }
                     });
         }
-    }
-
-    private void deleteDatafeedConfig(DeleteDatafeedAction.Request request, ActionListener<AcknowledgedResponse> listener) {
-        // Check datafeed is stopped
-        PersistentTasksCustomMetadata tasks = clusterService.state().getMetadata().custom(PersistentTasksCustomMetadata.TYPE);
-        if (MlTasks.getDatafeedTask(request.getDatafeedId(), tasks) != null) {
-            listener.onFailure(ExceptionsHelper.conflictStatusException(
-                    Messages.getMessage(Messages.DATAFEED_CANNOT_DELETE_IN_CURRENT_STATE, request.getDatafeedId(), DatafeedState.STARTED)));
-            return;
-        }
-
-        String datafeedId = request.getDatafeedId();
-
-        datafeedConfigProvider.getDatafeedConfig(
-            datafeedId,
-            ActionListener.wrap(
-                datafeedConfigBuilder -> {
-                    String jobId = datafeedConfigBuilder.build().getJobId();
-                    JobDataDeleter jobDataDeleter = new JobDataDeleter(client, jobId);
-                    jobDataDeleter.deleteDatafeedTimingStats(
-                        ActionListener.wrap(
-                            unused1 -> {
-                                datafeedConfigProvider.deleteDatafeedConfig(
-                                    datafeedId,
-                                    ActionListener.wrap(
-                                        unused2 -> listener.onResponse(AcknowledgedResponse.TRUE),
-                                        listener::onFailure));
-                            },
-                            listener::onFailure));
-                },
-                listener::onFailure));
     }
 
     @Override
